file update control, the newest files are compiled and dead .class files are purged, also all dependencies are only copied when the source file was changed

+++ src/main.rs
@@ -1,12 +1,13 @@
 mod config;
-mod maintenance;
 mod erros;
+mod maintenance;
 
-use erros::{ImportValidationError, BuildError, InitError};
-use std::{io::Write, path::PathBuf};
 use clap::{command, Parser, Subcommand};
 use config::ConfigFile;
-
+use erros::{BuildError, ImportValidationError, InitError};
+use std::collections::HashSet;
+use std::fs;
+use std::{io::Write, path::PathBuf};
 type Result<T> = std::result::Result<T, Box<dyn std::error::Error>>;
 
 #[derive(Parser)]
@@ -34,28 +35,22 @@ enum Target {
 fn main() {
     let args = Args::parse();
     match args.target {
-        Target::Build { jar } => {
-            match build(jar != 0) {
-                Err(e) => eprintln!("Build error: {}", e),
-                Ok(_) => println!("Build done"),
-            }
+        Target::Build { jar } => match build(jar != 0) {
+            Err(e) => eprintln!("Build error: {}", e),
+            Ok(_) => println!("Build done"),
         },
-        Target::Run { jar } => {
-            match run(jar != 0) {
-                Err(e) => eprintln!("Run error: {}", e),
-                Ok(_) => (),
-            }
+        Target::Run { jar } => match run(jar != 0) {
+            Err(e) => eprintln!("Run error: {}", e),
+            Ok(_) => (),
         },
         Target::Init { name } => {
             match init(name) {
                 Err(e) => eprintln!("Init error: {}", e),
                 Ok(_) => println!("Initialized project directory"),
             };
-        },
+        }
     };
 }
-
-
 
 fn build(jar: bool) -> Result<PostBuildData> {
     use std::process::Command;
@@ -66,7 +61,8 @@ fn build(jar: bool) -> Result<PostBuildData> {
 
     let working_dir = std::env::current_dir()?;
     let mut src_dir = working_dir.clone();
-    
+    let mut bin_dir = working_dir.clone();
+    bin_dir.push("bin");
     //gather external libs
     let mut external_libs: Vec<PathBuf> = vec![];
     if let Some(import) = &cfg.import {
@@ -74,7 +70,7 @@ fn build(jar: bool) -> Result<PostBuildData> {
             let path = PathBuf::from(&ext_lib.path);
             // let path = path.canonicalize()?;
             if path.extension().unwrap_or_default() != "jar" {
-                return Err(Box::new(ImportValidationError{path: path}))
+                return Err(Box::new(ImportValidationError { path: path }));
             }
             external_libs.push(path);
         }
@@ -112,38 +108,49 @@ fn build(jar: bool) -> Result<PostBuildData> {
     output.arg("-d");
     output.arg("bin");
 
-
     //gather sources
     src_dir.push("src");
     let sources = maintenance::get_sources(&src_dir, "java")?;
-    let sources = sources.into_iter()
+    let sources = sources
+        .into_iter()
         .map(|d| d.path())
         .collect::<Vec<_>>();
 
+    let mut source_count = 0;
     //pass recently modified sources
-    for (_index, src) in sources.iter()
-        .filter(|d| {
-            let Some(last_build) = &lockfile.last_build else {
-                return true;
-            };
-            match d.metadata() {
-                Err(_) => false,
-                Ok(m) => {
-                    let m = m.modified().unwrap();
-                    m.ge(last_build)
-                }
+    for src in sources.iter().filter(|d| {
+        let Some(last_build) = &lockfile.last_build else {
+            return true;
+        };
+
+        //check if the corresponding .class file already exists
+        let p = d.to_owned().clone();
+        let p = p.strip_prefix(&src_dir).unwrap().to_owned();
+        let mut p = bin_dir.join(p);
+        p.set_extension("class");
+        if !p.exists() {
+            return true;
+        }
+
+        match d.metadata() {
+            Err(_) => false,
+            Ok(m) => {
+                let m = m.modified().unwrap();
+                m.ge(last_build)
             }
-        })
-        .enumerate() {
+        }
+    }) {
         let source_arg = format!("{}", src.to_str().unwrap());
         //println!("adding source:`{}`", &source_arg);
         output.arg(source_arg);
+        source_count += 1;
     }
     //run build
-    let _output = output.output()?;
-    
-    std::io::stdout().write_all(&_output.stdout)?;
-    std::io::stderr().write_all(&_output.stderr)?;
+    if source_count > 0 {
+        let _output = output.output()?;
+        std::io::stdout().write_all(&_output.stdout)?;
+        std::io::stderr().write_all(&_output.stderr)?;
+    }
 
     //gather class files
     let classes = sources
@@ -153,9 +160,10 @@ fn build(jar: bool) -> Result<PostBuildData> {
             p.strip_prefix(&src_dir).unwrap().to_owned()
         })
         .collect::<Vec<PathBuf>>();
-    let mut bin_dir = working_dir.clone();
-    bin_dir.push("bin");
 
+    maintenance::purge_unused_classes(classes.clone())?;
+    
+    let old_lock = lockfile.clone();
     lockfile.last_build = Some(std::time::SystemTime::now());
     maintenance::create_lock_file(&lockfile)?;
 
@@ -168,10 +176,10 @@ fn build(jar: bool) -> Result<PostBuildData> {
         libs_arg: libs_arg,
         bin_dir: bin_dir,
         external_libs: external_libs,
-        current_lock: lockfile
+        current_lock: lockfile,
     };
     if jar {
-        jar_package(&post_build.cfg, &post_build)?;
+        package_jar(&post_build.cfg, &post_build, old_lock)?;
     }
 
     //make jar
@@ -179,11 +187,13 @@ fn build(jar: bool) -> Result<PostBuildData> {
     Ok(post_build)
 }
 
-
-
-fn jar_package(cfg: &ConfigFile, post_build: &PostBuildData) -> Result<()> {
-    use std::process::Command;
+fn package_jar(
+    cfg: &ConfigFile, 
+    post_build: &PostBuildData, 
+    lock: config::NopainLock) -> Result<()> 
+{
     use std::fs;
+    use std::process::Command;
 
     let package = cfg.package.name.as_str();
     let mut output = Command::new(&cfg.package.jar);
@@ -213,22 +223,40 @@ fn jar_package(cfg: &ConfigFile, post_build: &PostBuildData) -> Result<()> {
         .recursive(true)
         .create(&build_lib_dir)?;
 
-    for lib in post_build.libs.iter()
-        .map(|l| {
-            post_build.lib_dir.join(l).to_owned()
-        })
-        
+
+    //copy libs from /lib to /target/build/lib
+    for lib in post_build.libs
+        .iter()
+        .map(|l| post_build.lib_dir.join(l).to_owned())
     {
         let dest = build_lib_dir.join(lib.file_name().unwrap());
-        fs::copy(&lib, dest)?;
+        let meta = lib.metadata()?;
+
+        if let Some(last_build) = &lock.last_build {
+            if !dest.exists() || meta.modified()?.ge(last_build) {
+                fs::copy(&lib, dest)?;
+            }
+        } else {
+            if !dest.exists() {
+                fs::copy(&lib, dest)?;
+            }
+        }
     }
 
-    for ext in post_build.external_libs.iter()
-    {
+    for ext in post_build.external_libs.iter() {
         let dest = build_lib_dir.join(ext.file_name().unwrap());
-        fs::copy(&ext, dest)?;
+        let meta = ext.metadata()?;
+
+        if let Some(last_build) = &lock.last_build {
+            if !dest.exists() || meta.modified()?.ge(last_build) {
+                fs::copy(&ext, dest)?;
+            }
+        } else {
+            if !dest.exists() {
+                fs::copy(&ext, dest)?;
+            }
+        }
     }
-    
 
     std::io::stdout().write_all(&out.stdout)?;
     std::io::stderr().write_all(&out.stderr)?;
@@ -245,7 +273,7 @@ fn run(jar: bool) -> Result<()> {
             msg: "This package contains no entry point class".to_owned(),
         }));
     };
-    
+
     let mut output = Command::new(&cfg.package.java);
 
     // let class_files = build_data.classes.into_iter()
@@ -265,7 +293,6 @@ fn run(jar: bool) -> Result<()> {
         output.arg(main.as_str());
     }
 
-
     let _output = output.output()?;
 
     std::io::stdout().write_all(&_output.stdout)?;
@@ -273,8 +300,6 @@ fn run(jar: bool) -> Result<()> {
 
     Ok(())
 }
-
-
 
 fn init(name: String) -> Result<()> {
     use std::fs;
@@ -338,11 +363,15 @@ public class Main{
     Ok(())
 }
 
-fn generate_manifest(cfg: &ConfigFile, libs: &Vec<PathBuf>, external_libs: &Vec<PathBuf>) -> Result<PathBuf>{
+fn generate_manifest(
+    cfg: &ConfigFile,
+    libs: &Vec<PathBuf>,
+    external_libs: &Vec<PathBuf>,
+) -> Result<PathBuf> {
     use std::fs;
     let path = PathBuf::from("target/Manifest.txt");
     let mut manifest = fs::File::create(&path)?;
-    
+
     write!(manifest, "Manifest-Version: 1.0\n")?;
     if let Some(entry) = &cfg.package.main {
         write!(manifest, "Main-Class: {}\n", entry)?;
@@ -352,11 +381,17 @@ fn generate_manifest(cfg: &ConfigFile, libs: &Vec<PathBuf>, external_libs: &Vec<
         write!(manifest, "lib/{} ", lib.to_str().unwrap())?;
     }
     for ext in external_libs {
-        write!(manifest, "lib/{} ", ext.file_name().unwrap().to_str().unwrap())?;
+        write!(
+            manifest,
+            "lib/{} ",
+            ext.file_name().unwrap().to_str().unwrap()
+        )?;
     }
     write!(manifest, "\n\n")?;
     Ok(path)
 }
+
+
 struct PostBuildData {
     /// all .jar file paths used for compilation
     pub libs: Vec<PathBuf>,
@@ -371,5 +406,3 @@ struct PostBuildData {
     pub external_libs: Vec<PathBuf>,
     pub current_lock: config::NopainLock,
 }
-
-
